pyproject.toml
[project]
name = "migrol-coupon-code-generator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncio>=3.4.3",
    "python-barcode[images]>=0.15.1",
    "python-dotenv>=1.1.0",
    "python-telegram-bot>=22.0",
    "requests>=2.32.3",
]
